monitor/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
Flask==0.12.2
Flask-SocketIO==3.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
py-radix==0.10.0
pycparser==2.18
pyOpenSSL==18.0.0
python-engineio==2.2.0
python-socketio==1.7.6
requests==2.19.1
six==1.11.0
socketIO-client-nexus==0.7.6
urllib3==1.23
websocket-client==0.53.0
Werkzeug==0.14.1
